requirements.txt
jira==1.0.3
PyGithub==1.26.0
